// src/types/generated/graphql.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  timestamptz: any;
  uuid: any;
};
// ...
export type Match = {
   __typename?: 'match';
  created_at: Scalars['timestamptz'];
  date: Scalars['timestamptz'];
  game: Game;
  game_id: Scalars['Int'];
  ratings: Array<Rating>;
  ratings_aggregate: Rating_Aggregate;
  teams: Array<Team>;
  teams_aggregate: Team_Aggregate;
  updated_at: Scalars['timestamptz'];
  uuid: Scalars['uuid'];
};
// ...
export type Player = {
   __typename?: 'player';
  appearances: Array<Appearance>;
  appearances_aggregate: Appearance_Aggregate;
  created_at: Scalars['timestamptz'];
  id: Scalars['Int'];
  name: Scalars['String'];
  project: Project;
  project_id: Scalars['Int'];
  ratings: Array<Rating>;
  ratings_aggregate: Rating_Aggregate;
  updated_at: Scalars['timestamptz'];
};
// ...
export type Rating = {
   __typename?: 'rating';
  created_at: Scalars['timestamptz'];
  id: Scalars['Int'];
  match: Match;
  match_uuid: Scalars['uuid'];
  player: Player;
  player_id: Scalars['Int'];
  rating: Scalars['Int'];
  rating_before: Scalars['Int'];
  updated_at: Scalars['timestamptz'];
};
// ...
export type PlayerRatingByGameQueryVariables = {
  playerId: Scalars['Int'];
  gameId: Scalars['Int'];
  date: Scalars['timestamptz'];
};


export type PlayerRatingByGameQuery = (
  { __typename?: 'query_root' }
  & { rating: Array<(
    { __typename?: 'rating' }
    & Pick<Rating, 'rating'>
  )> }
);